dataproc leaveone

+++ utils.py
@@ -5,18 +5,23 @@ import os
 import numpy as np
 from collections import Counter
 
-def download_dataset(dataset_name, file_url, dir_path):
+def download_dataset(dataset_name, file_url, dataset_dir):
     '''
         数据集下载
     '''
+    # 检查是否存在源数据
+    if os.path.exists(dataset_dir):
+        return 
+
     print('\n==================================================【 %s 数据集下载】===================================================\n'%(dataset_name))
     print('url: %s\n'%(file_url))
 
+    dir_path = dataset_dir.split('/')[0]
     if dataset_name == 'UniMiB-SHAR' and file_url[-4:] == '.git': # 由于unimib数据集无法直接访问下载，这里我把unimib数据集上传到github进行访问clone
         if os.path.exists(os.path.join(dir_path, dataset_name)):
             shutil.rmtree(os.path.join(dir_path, dataset_name))
         os.system('git clone %s %s/%s' % (file_url, dir_path, dataset_name))
-        
+   
     else: # 其他数据集
         r = requests.get(url=file_url, stream=True)
         with open(os.path.join(dir_path, 'dataset.zip'),mode='wb') as f:  # 需要用wb模式
@@ -34,6 +39,10 @@ def download_dataset(dataset_name, file_url, dir_path):
         os.remove(os.path.join(dir_path, r'dataset.zip'))
 
     print()
+
+    # 检查数据集是否下载完毕
+    if not os.path.exists(dataset_dir):
+        quit('数据集下载失败，请检查url与网络后重试')
 
 
 def build_npydataset_readme(path):
@@ -59,3 +68,17 @@ def build_npydataset_readme(path):
             log = '\n===============================================================\n%s\n   x_train shape: %s\n   x_test shape: %s\n   y_train shape: %s\n   y_test shape: %s\n\n共【%d】个类别\ny_test中每个类别的样本数为 %s\n' % (dataset, x_train.shape, x_test.shape, y_train.shape, y_test.shape, category, new_d)
             w.write(log)
     os.chdir(curdir) # 返回原始地址
+
+def sliding_window(array, windowsize, overlaprate):
+    '''
+    array: 二维数据(n, m)，n为时序长度，m为模态轴数
+    windowsize: 窗口尺寸
+    overlaprate: 重叠率
+    '''
+    stride = int(windowsize * (1 - overlaprate)) # 计算stride
+    times = (len(array)-windowsize)//stride + 1 # 滑窗次数，同时也是滑窗后数据长度
+    res = []
+    for i in range(times):
+        x = array[i*stride : i*stride+windowsize]
+        res.append(x)
+    return res
